fix: clean some functions
