+++ frontend/src/layout/AuthenticationLayout.tsx
@@ -3,6 +3,5 @@ import { Navigate, Outlet } from "react-router";
 
 export const AuthenticationLayout = () => {
   const { isAuthenticated } = useAuthStore();
-  console.log("isAuthenticated", isAuthenticated);
   return isAuthenticated ? <Outlet /> : <Navigate to={"/"} replace />;
 };
