+++ frontend/src/utils/fetchInstance.ts
@@ -1,3 +1,4 @@
+// Todo: Will do some utils for reusable fetch fn...
 // type TFetchProps = {
 //   endpoint: string;
 //   options: Omit<RequestInit, "headers">;
